Updating migration precision

+++ db/migrate/20200314205358_create_waypoints.rb
@@ -1,8 +1,8 @@
 class CreateWaypoints < ActiveRecord::Migration[5.2]
   def change
     create_table :waypoints do |t|
-      t.decimal :latitude, precision: 16, scale: 14
-      t.decimal :longitude, precision: 16, scale: 14
+      t.decimal :latitude, precision: 17, scale: 14
+      t.decimal :longitude, precision: 17, scale: 14
       t.datetime :sent_at
       t.references :vehicle, foreign_key: true
